Tries to get working code

+++ user-service/src/main/java/ru/ct/belfort/db/UserEntity.java
@@ -4,5 +4,5 @@ package ru.ct.belfort.db;
 import lombok.Builder;
 
 @Builder
-public record UserEntity(long id, String str) {
+public record UserEntity(int id, String str) {
 }
